requirements/base.txt
Django==1.6.6
South==1.0
bpython==0.13.1
dj-static==0.0.6
django-authtools==1.0.0
django-braces==1.4.0
django-debug-toolbar==1.2.1
django-extensions==1.3.9
django-floppyforms==1.2.0
django-htmlmin==0.7.0
django-jsonview==0.4.3
django-model-utils==2.2
django-redis==3.7.1
docutils==0.12
logutils==0.3.3
psycopg2==2.5.3
pytz==2014.4
static==1.0.2
